+++ crates/zng-view/src/display_list.rs
@@ -11,14 +11,15 @@ use zng_view_api::{
     window::FrameId,
 };
 
-use crate::px_wr::PxToWr;
+use crate::{image_cache::ImageUseMap, px_wr::PxToWr};
 
 pub fn display_list_to_webrender(
     list: DisplayList,
     ext: &mut dyn DisplayListExtension,
+    images: &ImageUseMap,
     cache: &mut DisplayListCache,
 ) -> wr::BuiltDisplayList {
-    let r = display_list_build(&list, cache, ext, false);
+    let r = display_list_build(&list, cache, ext, images, false);
     cache.insert(list);
 
     r
@@ -286,6 +287,7 @@ impl DisplayListCache {
         mut end: usize,
         wr_list: &mut wr::DisplayListBuilder,
         ext: &mut dyn DisplayListExtension,
+        images: &ImageUseMap,
         sc: &mut SpaceAndClip,
     ) {
         if let Some(l) = self.lists.get(&frame_id) {
@@ -307,7 +309,7 @@ impl DisplayListCache {
                 &[]
             });
             for item in range {
-                display_item_to_webrender(item, wr_list, ext, sc, self, true);
+                display_item_to_webrender(item, wr_list, ext, images, sc, self, true);
             }
         } else {
             tracing::error!("did not find reuse frame {frame_id:?}");
@@ -347,9 +349,11 @@ impl DisplayListCache {
     /// Apply updates, returns the webrender update if the renderer can also be updated and there are any updates,
     /// or returns a new frame if a new frame must be rendered.
     #[expect(clippy::result_large_err)] // both are large
+    #[allow(clippy::too_many_arguments)]
     pub fn update(
         &mut self,
         ext: &mut dyn DisplayListExtension,
+        images: &ImageUseMap,
         transforms: Vec<FrameValueUpdate<PxTransform>>,
         floats: Vec<FrameValueUpdate<f32>>,
         colors: Vec<FrameValueUpdate<Rgba>>,
@@ -390,7 +394,7 @@ impl DisplayListCache {
         if new_frame {
             let list = self.lists.get_mut(&self.latest_frame).expect("no frame to update");
             let list = mem::take(&mut list.list);
-            let r = display_list_build(&list, self, ext, true);
+            let r = display_list_build(&list, self, ext, images, true);
             self.lists.get_mut(&self.latest_frame).unwrap().list = list;
 
             Err(r)
@@ -412,6 +416,7 @@ fn display_list_build(
     list: &[DisplayItem],
     cache: &mut DisplayListCache,
     ext: &mut dyn DisplayListExtension,
+    images: &ImageUseMap,
     is_reuse: bool,
 ) -> wr::BuiltDisplayList {
     let _s = tracing::trace_span!("DisplayList::build").entered();
@@ -424,7 +429,7 @@ fn display_list_build(
     });
 
     for item in list {
-        display_item_to_webrender(item, &mut wr_list, ext, &mut sc, cache, is_reuse);
+        display_item_to_webrender(item, &mut wr_list, ext, images, &mut sc, cache, is_reuse);
     }
 
     ext.display_list_end(&mut DisplayExtensionArgs {
@@ -439,6 +444,7 @@ fn display_item_to_webrender(
     item: &DisplayItem,
     wr_list: &mut wr::DisplayListBuilder,
     ext: &mut dyn DisplayListExtension,
+    images: &ImageUseMap,
     sc: &mut SpaceAndClip,
     cache: &DisplayListCache,
     is_reuse: bool,
@@ -449,7 +455,7 @@ fn display_item_to_webrender(
             seg_id,
             start,
             end,
-        } => cache.reuse(*frame_id, *seg_id, *start, *end, wr_list, ext, sc),
+        } => cache.reuse(*frame_id, *seg_id, *start, *end, wr_list, ext, images, sc),
 
         DisplayItem::PushReferenceFrame {
             id,
@@ -687,6 +693,11 @@ fn display_item_to_webrender(
                 spatial_id: sc.spatial_id(),
                 flags: sc.primitive_flags(),
             };
+
+            // images.
+            // !!: TODO get if image needs stripes to render, generate multiple push_image if it does
+            // ImageUseMap::new_use inserts the webrender::api::ImageData with the ExternalImageId
+            // It also contains the ImageTextureId with pixel size
 
             if tile_spacing.is_empty() && tile_size == image_size {
                 wr_list.push_image(