+++ crates/zng-view/src/image_cache/external.rs
@@ -9,6 +9,7 @@ use webrender::{
         units::{ImageDirtyRect, TexelRect},
     },
 };
+use zng_unit::{Px, PxRect, euclid};
 use zng_view_api::image::ImageTextureId;
 
 use super::{Image, ImageData};
@@ -38,10 +39,10 @@ impl ExternalImageHandler for WrImageCache {
         self.locked.push(img); // keep alive in case the image is removed mid-use
 
         match &**self.locked.last().unwrap() {
-            ImageData::RawData { pixels, .. } => {
+            ImageData::RawData { pixels, range, .. } => {
                 ExternalImage {
                     uv: TexelRect::invalid(), // `RawData` does not use `uv`.
-                    source: ExternalImageSource::RawData(&pixels[..]),
+                    source: ExternalImageSource::RawData(&pixels[range.clone()]),
                 }
             }
             ImageData::NativeTexture { uv, texture: id } => ExternalImage {
@@ -80,54 +81,184 @@ impl Image {
 /// The renderer must use [`WrImageCache`] as the external image source.
 #[derive(Default)]
 pub(crate) struct ImageUseMap {
-    id_tex: FxHashMap<ExternalImageId, (ImageTextureId, Image)>,
+    id_tex: FxHashMap<ExternalImageId, Box<[(ImageTextureId, Image)]>>,
     tex_id: FxHashMap<ImageTextureId, ExternalImageId>,
 }
 impl ImageUseMap {
     pub fn new_use(&mut self, image: &Image, document_id: DocumentId, api: &mut RenderApi) -> ImageTextureId {
         let id = image.external_id();
         match self.id_tex.entry(id) {
-            Entry::Occupied(e) => e.get().0,
+            Entry::Occupied(e) => e.get()[0].0,
             Entry::Vacant(e) => {
-                let key = api.generate_image_key();
-                let tex_id = ImageTextureId::from_raw(key.1);
-                e.insert((tex_id, image.clone())); // keep the image Arc alive, we expect this in `WrImageCache`.
-                self.tex_id.insert(tex_id, id);
+                if image.pixels().len() <= i32::MAX as usize {
+                    let key = api.generate_image_key();
+                    let tex_id = ImageTextureId::from_raw(key.1);
+                    e.insert(Box::new([(tex_id, image.clone())])); // keep the image Arc alive, we expect this in `WrImageCache`.
+                    self.tex_id.insert(tex_id, id);
 
-                let mut txn = webrender::Transaction::new();
-                txn.add_image(key, image.descriptor(), image.data(), None);
-                api.send_transaction(document_id, txn);
+                    let mut txn = webrender::Transaction::new();
+                    txn.add_image(key, image.descriptor(), image.data(), None);
+                    api.send_transaction(document_id, txn);
 
-                tex_id
+                    tex_id
+                } else {
+                    // Webrender uses i32 offsets for manipulating images, to support gigapixel images
+                    // we split the image into "stripes" that fit <=i32::MAX bytes
+
+                    let full_size = image.size();
+                    let w = full_size.width.0 as usize * 4;
+                    if w > i32::MAX as usize {
+                        tracing::error!("renderer does not support images with width * 4 > i32::MAX");
+                        return ImageTextureId::INVALID;
+                    }
+
+                    // find proportional split that fits, to avoid having the last stripe be to thin
+                    let full_height = full_size.height.0 as usize;
+                    let mut stripe_height = full_height / 2;
+                    while w * stripe_height > i32::MAX as usize {
+                        stripe_height /= 2;
+                    }
+                    let stripe_len = w * stripe_height;
+                    let stripes_len = full_height.div_ceil(stripe_height);
+                    let stripe_height = Px(stripe_height as _);
+
+                    // generate ImageTextureId for each stripe, we use the first stripe ID to identify in
+                    // the display list, when converted to Webrender display list the other stripes are injected
+                    let mut txn = webrender::Transaction::new();
+                    let mut stripes = Vec::with_capacity(stripes_len + 1);
+
+                    // we store the full image reference for `update_use`
+                    stripes.push((ImageTextureId::INVALID, image.clone()));
+
+                    // generate stripe images and transaction that associates each with a ImageTextureId
+                    for i in 0..stripes_len {
+                        let key = api.generate_image_key();
+                        let tex_id = ImageTextureId::from_raw(key.1);
+
+                        let y = stripe_height * Px(i as _);
+                        let mut size = full_size;
+                        size.height = stripe_height.min(full_size.height - y);
+                        let mut descriptor = image.descriptor();
+                        descriptor.size = euclid::size2(size.width.0, size.height.0);
+
+                        let offset = stripe_len * i;
+                        let range = offset..(stripe_len.min(image.pixels().len() - offset));
+
+                        let stripe = Image(Arc::new(ImageData::RawData {
+                            size,
+                            pixels: image.pixels().clone(),
+                            descriptor,
+                            density: image.density(),
+                            range,
+                        }));
+
+                        txn.add_image(key, stripe.descriptor(), stripe.data(), None);
+
+                        stripes.push((tex_id, stripe));
+                    }
+
+                    let tex_id = stripes[1].0;
+
+                    e.insert(stripes.into_boxed_slice());
+                    self.tex_id.insert(tex_id, id);
+
+                    api.send_transaction(document_id, txn);
+
+                    tex_id
+                }
             }
         }
     }
 
-    pub fn update_use(&mut self, texture_id: ImageTextureId, image: &Image, document_id: DocumentId, api: &mut RenderApi) {
+    pub fn update_use(
+        &mut self,
+        texture_id: ImageTextureId,
+        image: &Image,
+        dirty_rect: Option<PxRect>,
+        document_id: DocumentId,
+        api: &mut RenderApi,
+    ) -> bool {
         if let Entry::Occupied(mut e) = self.tex_id.entry(texture_id) {
             let id = image.external_id();
             if *e.get() != id {
-                let prev_id = e.insert(id);
-                self.id_tex.remove(&prev_id).unwrap();
-                self.id_tex.insert(id, (texture_id, image.clone()));
+                let prev_image = self.id_tex.get(&id).unwrap();
+                if prev_image[0].1.descriptor() != image.descriptor() {
+                    tracing::error!("cannot update image use {texture_id:?}, new image has different dimensions");
+                    return false;
+                }
 
+                let prev_id = e.insert(id);
+                let mut prev_image = self.id_tex.remove(&prev_id).unwrap();
                 let mut txn = webrender::Transaction::new();
-                txn.update_image(
-                    ImageKey(api.get_namespace_id(), texture_id.get()),
-                    image.descriptor(),
-                    image.data(),
-                    &ImageDirtyRect::All,
-                );
+
+                if prev_image.len() == 1 {
+                    prev_image[0] = (texture_id, image.clone());
+
+                    txn.update_image(
+                        ImageKey(api.get_namespace_id(), texture_id.get()),
+                        image.descriptor(),
+                        image.data(),
+                        &match dirty_rect {
+                            Some(r) => ImageDirtyRect::Partial(r.to_box2d().cast().cast_unit()),
+                            None => ImageDirtyRect::All,
+                        },
+                    );
+                } else {
+                    prev_image[0].1 = image.clone();
+                    let mut y = Px(0);
+                    for stripe in &mut prev_image[1..] {
+                        let img = ImageData::RawData {
+                            size: stripe.1.size(),
+                            pixels: image.pixels().clone(),
+                            descriptor: stripe.1.descriptor(),
+                            density: image.density(),
+                            range: stripe.1.range(),
+                        };
+                        stripe.1 = Image(Arc::new(img));
+
+                        txn.update_image(
+                            ImageKey(api.get_namespace_id(), stripe.0.get()),
+                            stripe.1.descriptor(),
+                            stripe.1.data(),
+                            &match dirty_rect {
+                                Some(mut r) => {
+                                    r.origin.y -= y;
+                                    let mut r = r.to_box2d();
+                                    r.min.y = r.min.y.max(Px(0));
+                                    r.max.y = r.max.y.min(stripe.1.size().height);
+
+                                    ImageDirtyRect::Partial(r.cast().cast_unit())
+                                }
+                                None => ImageDirtyRect::All,
+                            },
+                        );
+
+                        y += stripe.1.size().height;
+                    }
+                }
+
+                self.id_tex.insert(id, prev_image);
                 api.send_transaction(document_id, txn);
             }
+
+            true
+        } else {
+            tracing::error!("cannot update image use, texture not found");
+            false
         }
     }
 
     pub fn delete(&mut self, texture_id: ImageTextureId, document_id: DocumentId, api: &mut RenderApi) {
         if let Some(id) = self.tex_id.remove(&texture_id) {
-            let _img = self.id_tex.remove(&id); // remove but keep alive until the transaction is done.
+            let image = self.id_tex.remove(&id).unwrap(); // remove but keep alive until the transaction is done.
             let mut txn = webrender::Transaction::new();
-            txn.delete_image(ImageKey(api.get_namespace_id(), texture_id.get()));
+            if image.len() == 1 {
+                txn.delete_image(ImageKey(api.get_namespace_id(), image[0].0.get()));
+            } else {
+                for (texture_id, _) in &image[1..] {
+                    txn.delete_image(ImageKey(api.get_namespace_id(), texture_id.get()));
+                }
+            }
             api.send_transaction(document_id, txn);
         }
     }